Add a few CLI tests

+++ wof/wofpy_config.py
@@ -2,14 +2,12 @@ from __future__ import (absolute_import, division, print_function)
 
 import os
 import shutil
-
-import wof
+import sys
 
 from docopt import docopt
 
-__all__ = [
-    'main',
-    ]
+import wof
+
 
 __doc__ = """
 Generate configuration directory structure for running WOFpy.
@@ -40,17 +38,13 @@ _CONFIG = os.path.join(_ROOT, 'examples', 'production_configs')
 _ODM2_TIMESERIES = os.path.join(_ROOT, 'examples', 'flask', 'odm2', 'timeseries')
 
 
-def makedirs(directory, overwrite='soft'):
-    if not os.path.exists(directory):
-        os.makedirs(directory)
-    elif os.path.exists(directory) and overwrite in ['soft', 'hard']:
-        if overwrite == 'hard':
-            print('Overwriting directory {}'.format(directory))
-            shutil.rmtree(directory)
-            os.makedirs(directory)
-    else:
-        msg = 'Directory {} already exists.'.format
-        raise ValueError(msg(directory))
+def makedirs(directory, overwrite=None):
+    if os.path.exists(directory) and overwrite == 'soft':
+        return
+    elif os.path.exists(directory) and overwrite == 'hard':
+        sys.stdout.write('Overwriting directory {}\n'.format(directory))
+        shutil.rmtree(directory)
+    os.makedirs(directory)
 
 
 def copytree(src, dst, symlinks=False, ignore=None, overwrite='soft'):
@@ -60,7 +54,7 @@ def copytree(src, dst, symlinks=False, ignore=None, overwrite='soft'):
         s = os.path.join(src, item)
         d = os.path.join(dst, item)
         if overwrite == 'soft' and not os.path.exists(d):
-            print('Adding {}'.format(d))
+            sys.stdout.write('Adding {}\n'.format(d))
 
         if os.path.exists(d) and overwrite == 'soft':
             continue
